+++ tests/notion-oauth.test.js
@@ -5628,6 +5628,7 @@ function createWorkspaceVisualizationFixture(harness) {
                 'https://*.notion.so/*',
                 'https://github.com/*',
                 'https://www.github.com/*',
+                'https://gist.github.com/*',
                 'https://www.zhihu.com/*',
                 'https://zhuanlan.zhihu.com/*',
                 'http://*/*',